+++ FastAppFramework.Demo/App.xaml.cs
@@ -32,6 +32,7 @@ namespace FastAppFramework.Demo
             // Register types for navigation.
             containerRegistry.RegisterForNavigation<HomePage, HomePageViewModel>();
             containerRegistry.RegisterForNavigation<SamplePage, SamplePageViewModel>();
+            containerRegistry.RegisterForNavigation<ControlPage>();
             containerRegistry.RegisterForNavigation<DemoSettingsPage, DemoSettingsPageViewModel>();
         }
         protected override void RegisterSettingTypes(IApplicationSettingRegistry settingRegistry)